@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
   provider: ^6.0.1
@@ -31,8 +31,14 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  hive_flutter: ^1.1.0
+  hive: ^2.2.3
+  get_it: ^7.6.0
 
 dev_dependencies:
+  build_runner: ^2.3.3
+  hive_generator: ^2.0.0
+  flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
 
